pyproject.toml
[tool.poetry]
name = "aiohttp-task-server"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.5"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest-aiohttp = "^1.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
